feat(plasma-theme-builder,plasma-tokens-utils): Remove `transparent` tokens from `text` group
