+++ website/plasma-theme-builder/src/builder/themeTokenGetters/controlsSurfacesTokenGetters/getSurfaceTransparentAccentTokens.ts
@@ -1,6 +1,51 @@
-import type { TokensByType } from '@salutejs/plasma-tokens-utils';
+import type { ThemeConfig, TokensByType } from '@salutejs/plasma-tokens-utils';
 
-import { transparentAccentTokenCreator } from '../textIconsTokenGetters/getTextTransparentAccentTokens';
+export const transparentAccentTokenCreator = (comment: Record<keyof TokensByType, string>) => {
+    return (config: ThemeConfig) => {
+        const opacityDegrees = -0.88;
+        const accentColorDark = `${config.accentColor.dark}[${opacityDegrees}]`;
+        const accentColorLight = `${config.accentColor.light}[${opacityDegrees}]`;
+
+        return {
+            dark: {
+                default: {
+                    value: accentColorDark,
+                    comment: comment.default,
+                },
+                onDark: {
+                    value: accentColorDark,
+                    comment: comment.onDark,
+                },
+                onLight: {
+                    value: accentColorLight,
+                    comment: comment.onLight,
+                },
+                inverse: {
+                    value: accentColorLight,
+                    comment: comment.inverse,
+                },
+            },
+            light: {
+                default: {
+                    value: accentColorLight,
+                    comment: comment.default,
+                },
+                onDark: {
+                    value: accentColorDark,
+                    comment: comment.onDark,
+                },
+                onLight: {
+                    value: accentColorLight,
+                    comment: comment.onLight,
+                },
+                inverse: {
+                    value: accentColorDark,
+                    comment: comment.inverse,
+                },
+            },
+        };
+    };
+};
 
 const comment: Record<keyof TokensByType, string> = {
     default: 'Прозрачный акцентный фон поверхности/контрола',
